Initial commit. Basic functionallity tests pass.
db_driver includes the code for a brand new file based database. Each
thing has a folder, in that folder there are two jsons - reported.json
and desired.json, holding the current reported and desired AWS IOT
shadow equivalent. There is also a folder history, which has entries of
the sort reported.2017.txt, desired.2017.txt - an archived version of
the reported.json and desired.json, one per line. There could be another
folder in history, called archive, which stores histories more than 2
years old. It is updated on January 1st. The idea behind this homebrew
database is achieving less code, simple concepts (files and folders) and
easy observability by a stray admin with ssh access. This comes at the
cost of high specialization for the task - this is not a general purpose
database.

mqtt_operator losely captures the AWS IoT logic that we use with idiot.

Unit tests are written in gherkin style.