Adding mock thing, a python-based version of the C++ code. Aware of a desired race condition.
Almost every server communication, state maintenance function implemented, except software update.

The desired race condition occurs when desired is updated by GUI and then, before being sent to thing as an answer to it's message,
a thing reported update is received which overrides the new desired. Fix by adding a unused file which is created on gui update and deleted on answer to delta message.