Changing semantics of mode - mock thing now supports controlling gpios from server.
Changing semantics of mode. Mock thing now supports setting the mode of gpios, 0 1 or 'a'. This is the value thing will write on that gpio, except when 'a' - automatic, when actions can modify those values. If not specified, automatic is assumed.

Applying aliases to desired, discarding any user changes to them.
Instead of returning error when delta finds something in reported missing in desired, ignore it and let UI handle it in the future. It is non mission critical.