Portal handles, mock thing implements - time actions
Mock thing updated with the changes that need to be made to ESP to test the backend.

Decided to include two new fields in ESP state: 'b' which is boot time in unix time and
senses/time which is time of day as seen by the ESP.

Graph modified to not display the time sense.